+++ BookingService/src/main/java/com/dwellora/dto/UserDTO.java
@@ -1,105 +1,13 @@
-//package com.gl.app.dto;
-//
-//public class UserDTO {
-//
-//    private Integer userId;
-//
-//    private Integer apartmentId;
-//
-//    private String fullName;
-//
-//    private String email;
-//
-//    private String phone;
-//
-//    private String flatNumber;
-//
-//    private String role;
-//
-//    public UserDTO() {
-//    }
-//
-//    public UserDTO(Integer userId,
-//                   Integer apartmentId,
-//                   String fullName,
-//                   String email,
-//                   String phone,
-//                   String flatNumber,
-//                   String role) {
-//
-//        this.userId = userId;
-//        this.apartmentId = apartmentId;
-//        this.fullName = fullName;
-//        this.email = email;
-//        this.phone = phone;
-//        this.flatNumber = flatNumber;
-//        this.role = role;
-//    }
-//
-//    public Integer getUserId() {
-//        return userId;
-//    }
-//
-//    public void setUserId(Integer userId) {
-//        this.userId = userId;
-//    }
-//
-//    public Integer getApartmentId() {
-//        return apartmentId;
-//    }
-//
-//    public void setApartmentId(Integer apartmentId) {
-//        this.apartmentId = apartmentId;
-//    }
-//
-//    public String getFullName() {
-//        return fullName;
-//    }
-//
-//    public void setFullName(String fullName) {
-//        this.fullName = fullName;
-//    }
-//
-//    public String getEmail() {
-//        return email;
-//    }
-//
-//    public void setEmail(String email) {
-//        this.email = email;
-//    }
-//
-//    public String getPhone() {
-//        return phone;
-//    }
-//
-//    public void setPhone(String phone) {
-//        this.phone = phone;
-//    }
-//
-//    public String getFlatNumber() {
-//        return flatNumber;
-//    }
-//
-//    public void setFlatNumber(String flatNumber) {
-//        this.flatNumber = flatNumber;
-//    }
-//
-//    public String getRole() {
-//        return role;
-//    }
-//
-//    public void setRole(String role) {
-//        this.role = role;
-//    }
-//}
-
 package com.dwellora.dto;
 
+/**
+ * Data transfer object representing user details retrieved from the user service.
+ */
 public class UserDTO {
 
-    private Integer userId;
+    private Long userId;
 
-    private Integer apartmentId;
+    private Long apartmentId;
 
     private String fullName;
 
@@ -113,22 +21,21 @@ public class UserDTO {
 
     private String accountStatus;
 
-    public UserDTO() {
-    }
+    public UserDTO() {}
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getApartmentId() {
+    public Long getApartmentId() {
         return apartmentId;
     }
 
-    public void setApartmentId(Integer apartmentId) {
+    public void setApartmentId(Long apartmentId) {
         this.apartmentId = apartmentId;
     }
 
@@ -175,6 +82,7 @@ public class UserDTO {
     public String getAccountStatus() {
         return accountStatus;
     }
+
     public void setAccountStatus(String accountStatus) {
         this.accountStatus = accountStatus;
     }
